pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["suddenly*", "config*"]

[project]
name = "suddenly"
version = "0.1.0"
description = "Réseau fédéré de fiction partagée"
requires-python = ">=3.12"

dependencies = [
    "django>=5.0,<5.1",
    "gunicorn>=21.0",
    "whitenoise>=6.6",
    "psycopg[binary]>=3.1",
    "pydantic>=2.5",
    "django-allauth>=0.61",
    "django-htmx>=1.17",
    "djangorestframework>=3.14",
    "drf-spectacular>=0.27",
    "markdown>=3.5",
    "Pillow>=10.2",
]

[project.optional-dependencies]
federation = [
    "celery>=5.3",
    "redis>=5.0",
    "django-celery-beat>=2.5",
    "httpx>=0.27",
    "cryptography>=42.0",
    "django-ratelimit>=4.1",
]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.7",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pytest-playwright>=0.4",
    "factory-boy>=3.3",
    "ruff>=0.2",
    "mypy>=1.8",
    "django-stubs>=4.2",
    "django-extensions>=3.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main"]
exclude = ['migrations/']

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "allauth.*",
    "rest_framework.*",
    "django_ratelimit.*",
    "django_celery_beat.*",
    "factory.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short -m 'not e2e' --cov=suddenly --cov-fail-under=80"
markers = [
    "e2e: End-to-end browser tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
